crud education add

+++ src/app/components/education-section/education-section.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { FormArray, FormBuilder, Validators } from '@angular/forms';
 import { AutenticationService } from 'src/app/services/autentication.service';
 import { PortfolioService } from 'src/app/services/portfolio.service';
 import { Curso } from 'src/app/classes/curso';
@@ -14,30 +14,23 @@ export class EducationSectionComponent implements OnInit {
 
   selection:number = 0;
   miCurso = new Curso("","",[],"");
-  contador:number = 0;
+  contador!:number;
 
   formSwitchCrear:boolean = false;
   formSwitchUpdate:boolean = false;
-  form!:FormGroup;
   idCurso:number = 0;
-  arrayLenguajes:string[] = [];
-  arrayRandom:string[] = [];
-  values:string[] = [];
 
-  constructor(private autenticacionServicio:AutenticationService,private datosPortfolio:PortfolioService, private formBuilder:FormBuilder, private crudService:CrudService) {
-    this.form = formBuilder.group(
+  constructor(private autenticacionServicio:AutenticationService,private datosPortfolio:PortfolioService, private formBuilder:FormBuilder, private crudService:CrudService) {}
+
+    form = this.formBuilder.group(
       {
         nombre:["",[Validators.required,Validators.minLength(2)]],
         descripcion:["",[Validators.required,Validators.minLength(2)]],
         img:["",[Validators.required,Validators.minLength(2)]],
-        lenguajesAprendidos:new FormArray([
+        lenguajesAprendidos:this.formBuilder.array([
           this.formBuilder.control('')
         ])
       });
-  }
-
-    
-  
 
   arrayCursos:any;
 
@@ -52,11 +45,22 @@ export class EducationSectionComponent implements OnInit {
     })
   }
 
+  formInit(){
+    this.form = this.formBuilder.group(
+      {
+        nombre:["",[Validators.required,Validators.minLength(2)]],
+        descripcion:["",[Validators.required,Validators.minLength(2)]],
+        img:["",[Validators.required,Validators.minLength(2)]],
+        lenguajesAprendidos:this.formBuilder.array([
+          this.formBuilder.control('')
+        ])
+      });
+  }
+
   onClick(object:any){
     this.selection = object.id;
     this.miCurso.nombre = object.nombre;
     this.miCurso.descripcion = object.descripcion;
-    
     this.idCurso = object.id;
   }
 
@@ -69,6 +73,7 @@ export class EducationSectionComponent implements OnInit {
     var resetForm:HTMLFormElement;
   resetForm = <HTMLFormElement>document.getElementById('myForm');
     resetForm.reset();
+    this.formInit();
     this.contador = 0;
   }
   abrirFormUpdate(){
@@ -81,6 +86,7 @@ export class EducationSectionComponent implements OnInit {
   resetForm = <HTMLFormElement>document.getElementById('myForm');
   console.log(this.form);
     resetForm.reset();
+    this.formInit();
     this.contador = 0;
   }
   obtenerCurrentUserLength(){
@@ -90,7 +96,13 @@ export class EducationSectionComponent implements OnInit {
 
   construirCurso(){
 
-    let curso = new Curso("this.form.value.nombre", "this.form.value.descripcion", ["this.arrayRandom"], "this.form.value.img");
+    let nombre = this.form.get('nombre')?.value;
+    let descripcion = this.form.get('descripcion')?.value;
+    let arrayLenguajes = this.form.get(['lenguajesAprendidos'])?.value;
+    let img = this.form.get('img')?.value;
+
+    let curso = new Curso(nombre!, descripcion!, arrayLenguajes!, 
+      img!);
 
     return curso;
   }
@@ -110,12 +122,16 @@ export class EducationSectionComponent implements OnInit {
 
     this.cerrarFormCrear();
   }
+  get lenguajesAprendidos()
+  {
+    return this.form.get('lenguajesAprendidos') as FormArray;
+  }
 
-  anadirLenguaje(event:Event){
-    event.preventDefault;
-    this.contador++
-
+  anadirLenguaje(){
     this.lenguajesAprendidos.push(this.formBuilder.control(''));
+  }
+  quitarLenguaje(lenguajeIndex:number){
+    this.lenguajesAprendidos.removeAt(lenguajeIndex);
   }
   
   actualizarCurso(id:number, body:any){
@@ -129,18 +145,27 @@ export class EducationSectionComponent implements OnInit {
 
   updateCurso(event:Event){
     event.preventDefault;
-
+    this.contador = 0;
+    for(let leng of this.lenguajesAprendidos.controls)
+    {
+      console.log("lenguaje "+ this.form.get(['lenguajesAprendidos', this.contador])?.value);
+      console.log(this.form.get('nombre')?.value);
+      this.contador +=1;
+    }
     let curso = this.construirCurso();
+    
 
     console.log(curso);
     this.actualizarCurso(this.idCurso, curso);
-    // this.cerrarFormUpdate();
+    this.cerrarFormUpdate();
   }
   eliminarCurso(id:number){
     let path = "/eliminar/curso/";
-    this.crudService.deleteService(id, path).subscribe(()=>{
-      this.ngOnInit();
-    })
+    if(confirm("Are you sure?")){
+      this.crudService.deleteService(id, path).subscribe(()=>{
+        this.ngOnInit();
+      })
+    }
   }
 
 }
